[ISSUE #5036]♻️Refactor AuthenticationContext into trait and introduce BaseAuthenticationContext struct for extensible context management
