+++ rocketmq-auth/src/authentication/context/authentication_context.rs
@@ -15,75 +15,6 @@
 //  specific language governing permissions and limitations
 //  under the License.
 
-use std::any::Any;
-use std::collections::HashMap;
-
-use cheetah_string::CheetahString;
-
 use crate::authentication::AsAny;
 
-#[derive(Debug, Default)]
-pub struct AuthenticationContext {
-    channel_id: Option<CheetahString>,
-
-    rpc_code: Option<CheetahString>,
-
-    ext_info: HashMap<CheetahString, Box<dyn Any + Send + Sync>>,
-}
-
-impl AuthenticationContext {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn channel_id(&self) -> Option<&CheetahString> {
-        self.channel_id.as_ref()
-    }
-
-    pub fn set_channel_id(&mut self, channel_id: CheetahString) {
-        self.channel_id = Some(channel_id);
-    }
-
-    pub fn rpc_code(&self) -> Option<&CheetahString> {
-        self.rpc_code.as_ref()
-    }
-
-    pub fn set_rpc_code(&mut self, rpc_code: CheetahString) {
-        self.rpc_code = Some(rpc_code);
-    }
-}
-
-impl AuthenticationContext {
-    pub fn get_ext_info<T: 'static>(&self, key: &CheetahString) -> Option<&T> {
-        self.ext_info.get(key).and_then(|v| v.downcast_ref::<T>())
-    }
-    pub fn has_ext_info(&self, key: &CheetahString) -> bool {
-        self.ext_info.contains_key(key)
-    }
-    pub fn ext_info(&self) -> &HashMap<CheetahString, Box<dyn Any + Send + Sync>> {
-        &self.ext_info
-    }
-
-    pub fn ext_info_mut(&mut self) -> &mut HashMap<CheetahString, Box<dyn Any + Send + Sync>> {
-        &mut self.ext_info
-    }
-}
-
-impl AuthenticationContext {
-    pub fn set_ext_info<T>(&mut self, key: CheetahString, value: T)
-    where
-        T: Any + Send + Sync,
-    {
-        self.ext_info.insert(key, Box::new(value));
-    }
-}
-
-impl AsAny for AuthenticationContext {
-    fn as_any_mut(&mut self) -> &mut dyn Any {
-        self
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
+pub trait AuthenticationContext: AsAny {}
